+++ database/migrations/0001_01_01_000012_create_resep_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('id_barang');
             $table->unsignedBigInteger('id_bahan');
-            $table->integer('takaran');
+            $table->decimal('takaran');
             $table->softDeletes('deleted_at', precision:0);
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
